% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLmixed_add_functions.R
\name{coef.PLmod}
\alias{coef.PLmod}
\title{coef.PLmod}
\usage{
\method{coef}{PLmod}(object, ...)
}
\arguments{
\item{object}{an object of class PLmod}

\item{...}{Additional arguments from \code{\link{coef.merMod}}.}
}
\value{
sum of the random and fixed effects coefficients for each explanatory variable for
each level of the grouping factor.
}
\description{
Obtain coefficients for a model of class PLmod.
}
\keyword{GLLAMM}
\keyword{GLMM}
\keyword{IRT}
