% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLmixed_add_functions.R
\name{fitted.PLmod}
\alias{fitted.PLmod}
\title{fitted.PLmod}
\usage{
\method{fitted}{PLmod}(object, ...)
}
\arguments{
\item{object}{an object of class PLmod}

\item{...}{Additional arguments from \code{\link{fitted.merMod}}.}
}
\description{
Obtain fitted values for a model of class PLmod.
}
\keyword{GLLAMM}
\keyword{GLMM}
\keyword{IRT}
