% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLmixed_add_functions.R
\name{fixef.PLmod}
\alias{fixef.PLmod}
\title{fixef.PLmod}
\usage{
\method{fixef}{PLmod}(object, ...)
}
\arguments{
\item{object}{an object of class PLmod}

\item{...}{Additional arguments from \code{\link{fixef.merMod}}.}
}
\description{
Obtain fixed effect estimates for a model of class PLmod.
}
\keyword{GLLAMM}
\keyword{GLMM}
\keyword{IRT}
