% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLmixed_add_functions.R
\docType{data}
\name{IRTsim}
\alias{IRTsim}
\title{Simulated multilevel IRT dataset.}
\format{
A data frame with 2500 rows and 4 variables:
\describe{
  \item{sid}{Student ID}
  \item{school}{School ID}
  \item{item}{Item ID}
  \item{y}{Response}
}
}
\usage{
IRTsim
}
\description{
A simulated dataset that replicates the dataset from CITO.
}
\keyword{datasets}
