% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLmixed_add_functions.R
\name{iterPlot}
\alias{iterPlot}
\title{iterPlot}
\usage{
iterPlot(object)
}
\arguments{
\item{object}{an object of class PLmod}
}
\description{
Plot parameter estimates at each \code{\link{optim}} iteration.
}
\keyword{GLLAMM}
\keyword{GLMM}
\keyword{IRT}
