% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLmixed_add_functions.R
\docType{data}
\name{JUDGEsim}
\alias{JUDGEsim}
\title{Simulated Multi-rater Multi-response dataset.}
\format{
A data frame with 54462 rows and 7 variables:
\describe{
  \item{item}{Item ID}
  \item{method}{1 = teacher response, 2 = student response}
  \item{trait}{1 = trait 1, 2 = trait 2}
  \item{stu}{Student ID}
  \item{class}{Classroom ID}
  \item{tch}{Teacher ID}
  \item{response}{Item response}
}
}
\usage{
JUDGEsim
}
\description{
A simulated dataset that replicates the dataset from a multi-rater
mult-reponse study where teachers and students provided responses
about two student traits.
}
\keyword{datasets}
