% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLmixed_add_functions.R
\docType{data}
\name{KYPSitemsim}
\alias{KYPSitemsim}
\title{Simulated KYPS item-level dataset.}
\format{
A data frame with 66947 rows and 6 variables:
\describe{
  \item{id}{Student ID}
  \item{time}{Time Identifier}
  \item{item}{Item ID}
  \item{mid}{Middle School ID}
  \item{hid}{High School ID}
  \item{response}{Item Response}
}
}
\usage{
KYPSitemsim
}
\description{
A simulated dataset that replicates the dataset item-level
data from KYPS.
}
\keyword{datasets}
