% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLmixed_add_functions.R
\docType{data}
\name{KYPSsim}
\alias{KYPSsim}
\title{Simulated KYPS dataset.}
\format{
A data frame with 11494 rows and 5 variables:
\describe{
  \item{mid}{Middle School ID}
  \item{hid}{High School ID}
  \item{sid}{Student ID}
  \item{time}{Time Identifier}
  \item{esteem}{Self Esteem}
}
}
\usage{
KYPSsim
}
\description{
A simulated dataset that replicates the dataset from KYPS.
}
\keyword{datasets}
