% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLmixed_main_function.R
\docType{package}
\name{PLmixed-package}
\alias{PLmixed-package}
\title{PLmixed: A package for estimating GLMMs with factor structures.}
\description{
The \code{PLmixed} package's main function is \code{\link{PLmixed}}, which estimates
the model through nested maximizations using the \pkg{\link{lme4}} package
and \code{\link{optim}} function. This extends the capabilities of \pkg{\link{lme4}}
to allow for estimated factor structures, making it useful for estimating multilevel
factor analysis and item response theory models with an arbitrary number of hierarchical
levels or crossed random effects.
}
