% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLmixed_add_functions.R
\name{plot.PLmod}
\alias{plot.PLmod}
\title{plot.PLmod}
\usage{
\method{plot}{PLmod}(x, ...)
}
\arguments{
\item{x}{an object of class PLmod}

\item{...}{Additional arguments from \code{\link{plot.merMod}}.}
}
\description{
Diagnostic plots for a model of class PLmod.
}
\keyword{GLLAMM}
\keyword{GLMM}
\keyword{IRT}
