% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLmixed_add_functions.R
\name{predict.PLmod}
\alias{predict.PLmod}
\title{predict.PLmod}
\usage{
\method{predict}{PLmod}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{an object of class PLmod}

\item{newdata}{data frame to obtain predictions for}

\item{...}{Additional arguments from \code{\link{predict.merMod}}.}
}
\description{
Predict response values from a model of class PLmod.
}
\keyword{GLLAMM}
\keyword{GLMM}
\keyword{IRT}
