% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLmixed_add_functions.R
\name{print.summary.PLmod}
\alias{print.summary.PLmod}
\title{print.summary.PLmod}
\usage{
\method{print}{summary.PLmod}(x, digits = 4, ...)
}
\arguments{
\item{x}{an object of class PLmod}

\item{digits}{minimal number of significant digits, see \code{\link{print.default}}.}

\item{...}{Additional arguments.}
}
\description{
Print the output for a PLmixed model object of class PLmod.
}
\keyword{GLLAMM}
\keyword{GLMM}
\keyword{IRT}
