% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLmixed_add_functions.R
\name{ranef.PLmod}
\alias{ranef.PLmod}
\title{ranef.PLmod}
\usage{
\method{ranef}{PLmod}(object, ...)
}
\arguments{
\item{object}{an object of class PLmod}

\item{...}{Additional arguments from \code{\link{ranef.merMod}}.}
}
\description{
Obtain conditional modes of the random effects for a model of class PLmod.
}
\keyword{GLLAMM}
\keyword{GLMM}
\keyword{IRT}
