% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLmixed_add_functions.R
\name{residuals.PLmod}
\alias{residuals.PLmod}
\title{residuals.PLmod}
\usage{
\method{residuals}{PLmod}(object, ...)
}
\arguments{
\item{object}{an object of class PLmod}

\item{...}{Additional arguments from \code{\link{residuals.merMod}}.}
}
\description{
Obtain residuals for a model of class PLmod.
}
\keyword{GLLAMM}
\keyword{GLMM}
\keyword{IRT}
