% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLmixed_add_functions.R
\name{simulate.PLmod}
\alias{simulate.PLmod}
\title{simulate.PLmod}
\usage{
\method{simulate}{PLmod}(object, ...)
}
\arguments{
\item{object}{an object of class PLmod}

\item{...}{Additional arguments from \code{\link{simulate.merMod}}.}
}
\description{
Simulate responses from a model of class PLmod.
}
\keyword{GLLAMM}
\keyword{GLMM}
\keyword{IRT}
