% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLmixed_add_functions.R
\name{summary.PLmod}
\alias{summary.PLmod}
\title{summary.PLmod}
\usage{
\method{summary}{PLmod}(object, ...)
}
\arguments{
\item{object}{an object of class PLmod}

\item{...}{Additional arguments.}
}
\value{
An object containing all parameter estimates and model characteristics.
}
\description{
Obtain key output for a fitted PLmixed model object of class PLmod.
}
\keyword{GLLAMM}
\keyword{GLMM}
\keyword{IRT}
